tax_calc/requirements/base.txt
Django==3.2.4
djangorestframework==3.12.2
psycopg2==2.8.6
